# chore: parallelise inverse polynomial construction for lookup relations