--- a/barretenberg/cpp/src/barretenberg/relations/logderiv_lookup_relation.hpp
+++ b/barretenberg/cpp/src/barretenberg/relations/logderiv_lookup_relation.hpp
@@ -3,6 +3,7 @@
 #include <tuple>
 
 #include "barretenberg/common/constexpr_utils.hpp"
+#include "barretenberg/common/thread.hpp"
 #include "barretenberg/honk/proof_system/logderivative_library.hpp"
 #include "barretenberg/polynomials/univariate.hpp"
 #include "barretenberg/relations/relation_types.hpp"
@@ -157,16 +158,25 @@ template <typename FF_> class LogDerivLookupRelationImpl {
     {
         auto& inverse_polynomial = get_inverse_polynomial(polynomials);
 
-        for (size_t i = 0; i < circuit_size; ++i) {
-            // We only compute the inverse if this row contains a lookup gate or data that has been looked up
-            if (polynomials.q_lookup.get(i) == 1 || polynomials.lookup_read_tags.get(i) == 1) {
-                // TODO(https://github.com/AztecProtocol/barretenberg/issues/940): avoid get_row if possible.
-                auto row = polynomials.get_row(i); // Note: this is a copy. use sparingly!
-                auto value = compute_read_term<FF, 0>(row, relation_parameters) *
-                             compute_write_term<FF, 0>(row, relation_parameters);
-                inverse_polynomial.at(i) = value;
+        size_t min_iterations_per_thread = 1 << 6; // min number of iterations for which we'll spin up a unique thread
+        size_t num_threads = bb::calculate_num_threads_pow2(circuit_size, min_iterations_per_thread);
+        size_t iterations_per_thread = circuit_size / num_threads; // actual iterations per thread
+
+        parallel_for(num_threads, [&](size_t thread_idx) {
+            size_t start = thread_idx * iterations_per_thread;
+            size_t end = (thread_idx + 1) * iterations_per_thread;
+            for (size_t i = start; i < end; ++i) {
+                // We only compute the inverse if this row contains a lookup gate or data that has been looked up
+                if (polynomials.q_lookup.get(i) == 1 || polynomials.lookup_read_tags.get(i) == 1) {
+                    // TODO(https://github.com/AztecProtocol/barretenberg/issues/940): avoid get_row if possible.
+                    auto row = polynomials.get_row(i); // Note: this is a copy. use sparingly!
+                    auto value = compute_read_term<FF, 0>(row, relation_parameters) *
+                                 compute_write_term<FF, 0>(row, relation_parameters);
+                    inverse_polynomial.at(i) = value;
+                }
             }
-        }
+        });
+
         // Compute inverse polynomial I in place by inverting the product at each row
         FF::batch_invert(inverse_polynomial.coeffs());
     };
@@ -233,15 +243,14 @@ template <typename FF_> class LogDerivLookupRelationImpl {
         const auto write_inverse = inverses * read_term;                        // Degree 3 (4)
         const auto read_inverse = inverses * write_term;                        // Degree 2 (3)
 
-        // Establish the correctness of the polynomial of inverses I. Note: inverses is computed so that the value is 0
-        // if !inverse_exists.
-        // Degrees:                     2 (3)       1 (2)        1              1
+        // Establish the correctness of the polynomial of inverses I. Note: inverses is computed so that the value
+        // is 0 if !inverse_exists. Degrees:                     2 (3)       1 (2)        1              1
         std::get<0>(accumulator) +=
             ShortView((read_term * write_term * inverses - inverse_exists) * scaling_factor); // Deg 4 (6)
 
-        // Establish validity of the read. Note: no scaling factor here since this constraint is 'linearly dependent,
-        // i.e. enforced across the entire trace, not on a per-row basis.
-        // Degrees:                       1            2 (3)            1            3 (4)
+        // Establish validity of the read. Note: no scaling factor here since this constraint is 'linearly
+        // dependent, i.e. enforced across the entire trace, not on a per-row basis. Degrees: 1            2 (3) 1
+        // 3 (4)
         std::get<1>(accumulator) += read_selector * read_inverse - read_counts * write_inverse; // Deg 4 (5)
     }
 };