Benchmark were showing that oink is the second most expensive round in PG after combiner. On top of that one component where we see discrepancies when increasing the ambient trace size is logderivative inverses construction. A step towards improving this is parallelising the construction of inverse polynomials (which is linear). Also the inverses can be committed to with `commit_sparse` which shows a slight improvement as well.
BEFORE
```
CLIENT_IVC_BENCH_STRUCTURE(2^19)

ClientIVCBench/Full/6      29146 ms        27299 ms
ProtogalaxyProver::prove(t)            16265    58.29%
ProtogalaxyProver_::run_oink_prover_on_each_incomplete_key(t)    5624    34.58%


EXAMPLE_20(2^20)

ClientIVCBench/Full/6      37145 ms        34235 ms
ProtogalaxyProver::prove(t)            21283    60.75%
ProtogalaxyProver_::run_oink_prover_on_each_incomplete_key(t)    8818    41.43%
COMMIT::lookup_inverses(t)        406     9.82%
```

AFTER
```
CLIENT_IVC_BENCH_STRUCTURE(2^19)

ClientIVCBench/Full/6      27351 ms        25477 ms 
ProtogalaxyProver::prove(t)            14627    55.72%
ProtogalaxyProver_::run_oink_prover_on_each_incomplete_key(t)    4030    27.55%


EXAMPLE_20(2^20)
ClientIVCBench/Full/6      33852 ms        30893 ms   
ProtogalaxyProver::prove(t)            18250    56.97%
ProtogalaxyProver_::run_oink_prover_on_each_incomplete_key(t)    5526    30.28%
COMMIT::lookup_inverses(t)        301     7.43%
```